requirements.txt
streamlit>=1.28.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0